chore(typescript): converted to TS
